First pass at splitting up types into separate files
