+++ ical/freebusy.py
@@ -10,8 +10,8 @@ from typing import Any, Optional, Union
 
 from pydantic import Field, validator
 
+from ._types import CalAddress, ComponentModel, Period, RequestStatus, Uri
 from .parsing.property import ParsedProperty
-from .types import CalAddress, ComponentModel, Period, RequestStatus, Uri
 from .util import dtstamp_factory, normalize_datetime, uid_factory
 
 _LOGGER = logging.getLogger(__name__)
