+++ ical/types/__init__.py
@@ -1,0 +1,1 @@
+"""Library for parsing rfc5545 types."""
