+++ tests/test_freebusy.py
@@ -10,8 +10,9 @@ from unittest.mock import patch
 import pytest
 from pydantic import ValidationError
 
+from ical._types import Period
 from ical.freebusy import FreeBusy
-from ical.types import FreeBusyType, Period
+from ical.types.const import FreeBusyType
 
 
 @pytest.fixture(autouse=True)
