+++ tests/test_types.py
@@ -6,42 +6,9 @@ from typing import Optional, Union
 import pytest
 from pydantic import ValidationError
 
+from ical._types import ICS_ENCODERS, ComponentModel, Period, Priority, UtcOffset
 from ical.parsing.component import ParsedComponent
-from ical.parsing.property import ParsedProperty, ParsedPropertyParameter
-from ical.types import ICS_ENCODERS, ComponentModel, Geo, Period, Priority, UtcOffset
-from ical.tzif import timezoneinfo
-
-
-def test_text() -> None:
-    """Test for a text property value."""
-
-    class TestModel(ComponentModel):
-        """Model with a Text value."""
-
-        text_value: str
-
-    component = ParsedComponent(name="text-model")
-    component.properties.append(
-        ParsedProperty(
-            name="text_value",
-            value="Project XYZ Final Review\\nConference Room - 3B\\nCome Prepared.",
-        )
-    )
-    model = TestModel.parse_obj(component.as_dict())
-    assert model == {
-        "text_value": "\n".join(
-            ["Project XYZ Final Review", "Conference Room - 3B", "Come Prepared."]
-        )
-    }
-    assert model.__encode_component_root__() == ParsedComponent(
-        name="TestModel",
-        properties=[
-            ParsedProperty(
-                name="text_value",
-                value="Project XYZ Final Review\\nConference Room - 3B\\nCome Prepared.",
-            )
-        ],
-    )
+from ical.parsing.property import ParsedProperty
 
 
 def test_encode_component() -> None:
@@ -111,194 +78,6 @@ def test_encode_component() -> None:
             ],
         ),
     ]
-
-
-def test_datedatime_parser() -> None:
-    """Test for a datetime property value."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        dt: datetime.datetime
-
-    model = TestModel.parse_obj(
-        {
-            "dt": [ParsedProperty(name="dt", value="20220724T120000")],
-        }
-    )
-    assert model.dt == datetime.datetime(2022, 7, 24, 12, 0, 0)
-
-    # Build from an object
-    model = TestModel(dt=datetime.datetime(2022, 7, 20, 13, 0, 0))
-    assert model.dt.isoformat() == "2022-07-20T13:00:00"
-
-
-def test_datedatime_value_parser() -> None:
-    """Test a datetime with a property parameter value."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        dt: Union[datetime.datetime, datetime.date]
-
-    model = TestModel.parse_obj(
-        {
-            "dt": [
-                ParsedProperty(
-                    name="dt",
-                    value="20220724T120000",
-                    params=[
-                        ParsedPropertyParameter(name="VALUE", values=["DATE-TIME"]),
-                    ],
-                )
-            ],
-        }
-    )
-    assert model.dt == datetime.datetime(2022, 7, 24, 12, 0, 0)
-    model = TestModel.parse_obj(
-        {
-            "dt": [
-                ParsedProperty(
-                    name="dt",
-                    value="20220724",
-                    params=[
-                        ParsedPropertyParameter(name="VALUE", values=["DATE"]),
-                    ],
-                )
-            ],
-        }
-    )
-    assert model.dt == datetime.date(2022, 7, 24)
-
-    with pytest.raises(ValidationError):
-        TestModel.parse_obj(
-            {
-                "dt": [
-                    ParsedProperty(
-                        name="dt",
-                        value="20220724T120000",
-                        params=[
-                            ParsedPropertyParameter(name="VALUE", values=["INVALID"]),
-                        ],
-                    )
-                ],
-            }
-        )
-
-    with pytest.raises(ValidationError):
-        TestModel.parse_obj(
-            {
-                "dt": [
-                    ParsedProperty(
-                        name="dt",
-                        value="20220724",
-                        params=[
-                            ParsedPropertyParameter(name="VALUE", values=["INVALID"]),
-                        ],
-                    )
-                ],
-            }
-        )
-
-
-def test_datedatime_parameter_encoder() -> None:
-    """Test a datetime with a property parameter encoded in the output."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        dt: datetime.datetime
-
-        class Config:
-            """Pydantic model configuration."""
-
-            json_encoders = ICS_ENCODERS
-
-    model = TestModel.parse_obj(
-        {
-            "dt": [
-                ParsedProperty(
-                    name="dt",
-                    value="20220724T120000",
-                    params=[
-                        ParsedPropertyParameter(
-                            name="TZID", values=["America/New_York"]
-                        ),
-                    ],
-                )
-            ],
-        }
-    )
-    assert model.dt == datetime.datetime(
-        2022, 7, 24, 12, 0, 0, tzinfo=timezoneinfo.read_tzinfo("America/New_York")
-    )
-    assert model.__encode_component_root__() == ParsedComponent(
-        name="TestModel",
-        properties=[
-            ParsedProperty(
-                name="dt",
-                value="20220724T120000",
-                params=[
-                    ParsedPropertyParameter(name="TZID", values=["America/New_York"])
-                ],
-            )
-        ],
-    )
-
-
-def test_date_parser() -> None:
-    """Test for a date property value."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        d: datetime.date
-
-    model = TestModel.parse_obj(
-        {
-            "d": [ParsedProperty(name="d", value="20220724")],
-        }
-    )
-    assert model.d == datetime.date(2022, 7, 24)
-
-    with pytest.raises(ValidationError):
-        TestModel.parse_obj(
-            {
-                "d": [
-                    ParsedProperty(
-                        name="dt",
-                        value="invalid-value",
-                    )
-                ],
-            }
-        )
-
-    # Build from an object
-    model = TestModel(d=datetime.date(2022, 7, 20))
-    assert model.d.isoformat() == "2022-07-20"
-
-
-def test_union_date_parser() -> None:
-    """Test for a union of multiple date property values."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        d: Union[datetime.datetime, datetime.date]
-
-    model = TestModel.parse_obj(
-        {
-            "d": [ParsedProperty(name="d", value="20220724")],
-        }
-    )
-    assert model.d == datetime.date(2022, 7, 24)
-
-    model = TestModel.parse_obj(
-        {
-            "d": [ParsedProperty(name="d", value="20220724T120000")],
-        }
-    )
-    assert model.d == datetime.datetime(2022, 7, 24, 12, 0, 0)
 
 
 def test_list_parser() -> None:
@@ -472,24 +251,6 @@ def test_duration_from_object() -> None:
     component = model.__encode_component_root__()
     assert component.name == "TestModel"
     assert component.properties == [ParsedProperty(name="duration", value="PT1H")]
-
-
-def test_geo() -> None:
-    """Test for geo fields."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        geo: Geo
-
-    model = TestModel.parse_obj(
-        {"geo": [ParsedProperty(name="geo", value="120.0;-30.1")]}
-    )
-    assert model.geo.lat == 120.0
-    assert model.geo.lng == -30.1
-
-    with pytest.raises(ValidationError):
-        TestModel.parse_obj({"geo": [ParsedProperty(name="geo", value="10")]})
 
 
 def test_integer() -> None:
